add whitespace to code for readability and move MessageList to ChatWindow
